1.1

+++ 0x14-bit_manipulation/0-binary_to_uint.c
@@ -12,7 +12,7 @@
  * Return: the answer
  */
 
-unsigned int _pow_recursion(int x, int y)
+unsigned long int _pow_recursion(int x, int y)
 {
 	unsigned int sum;
